+++ src/app.html
@@ -36,4 +36,5 @@
   <body data-sveltekit-preload-data="hover">
     <div style="display: contents">%sveltekit.body%</div>
   </body>
+  <footer />
 </html>
