--- src/lib/ui/News.svelte
@@ -1,131 +1,0 @@
-<section class="column is-12 mt-6">
-    <section class="p-4">
-      <div class="has-text-centered pt-6 ">
-        <p class="title has-text-weight-bold">News
-        </p>
-        <p class="subtitle">
-          Every days new for travelers
-        </p>
-      </div>
-    </section>
-    <article class="content">
-      <div class="container">
-        <main class="columns content mt-4">
-          <section class="column is-full px-6" >
-            <h3> Latest News </h3>
-            <hr>
-          </section>
-        </main>
-        <div class="columns is-multiline">
-          <div class="column is-9 is-9-desktop">
-            <section class="columns is-multiline box mt-3">
-              <div class="column is-3">
-                <div class="mb-6 is-flex">
-                  <figure class="image is-256x256 m-auto">
-                    <a href="https://ibb.co/QFYPbTX">
-                    <img src="https://i.ibb.co/Lz9xkcv/city-destination.jpg" alt="City destination" class="pt-3">
-                    </a>
-                  </figure>
-                </div>
-              </div>
-              <div class="column is-9">
-                <div class="ml-3 has-text-grey ">
-                  <h4 class="is-size-4 has-text-weight-bold mb-2 pb-3">City destinations</h4>
-                  <p>
-                    If you love Venice, you might want to read this article and find out about other beautiful cities with canals going through.
-                  </p>
-                  <p><a href="/" class="has-text-grey-light">Read more</a></p>
-                </div>
-              </div>
-            </section>
-            <section class="columns is-multiline box mt-3">
-              <div class="column is-3">
-                <div class="mb-6 is-flex">
-                  <figure class="image is-256x256 m-auto">
-                    <a href="https://ibb.co/ZTR1Yx5">
-                    <img src="https://i.ibb.co/q1LFWpK/new-zealand.jpg" alt="New Zealand" class="pt-3">
-                    </a>
-                  </figure>
-                </div>
-              </div>
-              <div class="column is-9">
-                <div class="ml-3 has-text-grey ">
-                  <h4 class="is-size-4 has-text-weight-bold mb-2 pb-3">Exotic destinations, why you should go</h4>
-                  <p>
-                    Plenty of unexpected and exciting things are to happen once you embark upon a journey on these exotic destinations. Let them flip you out!
-                  </p>
-                  <p><a href="/" class="has-text-grey-light">Read more</a></p>
-                </div>
-              </div>
-            </section>
-            <section class="columns is-multiline box mt-3">
-              <div class="column is-3">
-                <div class="mb-6 is-flex">
-                  <figure class="image is-256x256 m-auto ">
-                    <a href="https://ibb.co/rc4rMDQ">
-                    <img src="https://i.ibb.co/s6yYQ8g/city-tourism.jpg" alt="City tourism" class="pt-3">
-                    </a>
-                  </figure>
-                </div>
-              </div>
-              <div class="column is-9">
-                <div class="ml-3 has-text-grey ">
-                  <h4 class="is-size-4 has-text-weight-bold mb-2 pb-3">Tourist events near you</h4>
-                  <p>
-                    Discover all the tourist events near to you with our Geolocating Tourist Events APP that you can download to your phone.
-                  </p>
-                  <p><a href="/" class="has-text-grey-light">Read more</a></p>
-                </div>
-              </div>
-            </section>
-            <section class="columns is-multiline box mt-3">
-              <div class="column is-3">
-                <div class="mb-6 is-flex">
-                  <figure class="image is-256x256 m-auto">
-                    <a href="https://ibb.co/QNMzJNF">
-                    <img src="https://i.ibb.co/R0cZg0p/exotic-beach.jpg" alt="Exotic beach" class="pt-3">
-                    </a>
-                  </figure>
-                </div>
-              </div>
-              <div class="column is-9">
-                <div class="ml-3 has-text-grey ">
-                  <h4 class="is-size-4 has-text-weight-bold mb-2 pb-3">Tropical destinations? Why not?</h4>
-                  <p>
-                    Whether you are a seasoned traveler or a first-time one, take a look at our traveling guide.
-                  </p>
-                  <p><a href="/" class="has-text-grey-light">Read more</a>
-                  </p>
-                </div>
-              </div>
-            </section>
-          </div>
-          <div class="column content is-3 ">
-            <aside class="content box has-text-grey">
-              <h2 class="has-text-info">Weekly Deals</h2>
-              <h3>Our travel destinations</h3>
-              <figure class="image">
-                <a href="https://ibb.co/QFYPbTX">
-                <img src="https://i.ibb.co/Lz9xkcv/city-destination.jpg" alt="City destination" class="is-fullwidth">
-                </a>
-              </figure>
-              <p class="p-3">
-                Here comes the next promos for Summer 2024. This time we are gravitating towards cultural tourism and destinations. Why don't you take a quick peek of what is in store for you?<br> 
-                <a href="/" class="has-text-grey-light">Read more</a>
-              </p>
-              <h3>Insanity Deals</h3>
-              <figure class="image">
-                <a href="https://ibb.co/ZTR1Yx5">
-                <img src="https://i.ibb.co/q1LFWpK/new-zealand.jpg" alt="New Zealand" class="is-fullwidth">
-                </a>
-              </figure>
-              <p class="p-3">
-                With the new exotic destinations added to our promos, hurry up as we have new <a href="/" class="has-text-grey-light">Insanity Deals</a> that are worth checking this week.
-                Find out more about this irrepeatable deals and do not forget to give us your feedback when you are back in town ;-) !
-              </p>
-            </aside>
-          </div>
-        </div>
-      </div>
-    </article>
-  </section>
