+++ src/routes/+layout.svelte
@@ -1,0 +1,45 @@
+<!-- <script >
+    import { loggedInUser } from "$lib/runes.svelte";
+    import Heading from "$lib/ui/Heading.svelte";
+    import Menu from "$lib/ui/Menu.svelte";
+  </script>
+  
+  <div class="container">
+    {#if loggedInUser.email}
+      <Menu />
+      <Heading />
+    {/if}
+    <slot />
+  </div> -->
+  
+  
+
+  <script lang="ts" context="module">
+    // https://www.npmjs.com/package/svelte-fa
+    import Footer from "$lib/Footer.svelte";
+    import Header from "$lib/Header.svelte";
+    import WelcomeMenu from "$lib/ui/WelcomeMenu.svelte";
+    import SplashScreen from "$lib/ui/SplashScreen.svelte";
+  
+    // @ts-ignore
+    export const load = async ({ page }) => ({
+      props: {
+        key: page.path
+      }
+    });
+  
+    import SignupForm from "./signup/SignupForm.svelte";
+    import { fromStore } from "svelte/store";
+    import Homepage from "$lib/ui/Homepage.svelte";
+    /**
+     * @type {any}
+     */
+    export let key: any;
+  </script>
+  
+  <WelcomeMenu />
+  <div class="container">
+    <!-- <SignupForm /> -->
+    <slot />
+  </div>
+  <Footer />
