+++ src/routes/signup/+page.svelte
@@ -1,28 +1,25 @@
 <script context="module">
-    // https://www.npmjs.com/package/svelte-fa
-    import Footer from "$lib/Footer.svelte";
-    import Header from "$lib/Header.svelte"; 
-    import SignupForm from "$lib/ui/SignupForm.svelte";
-  
-      // @ts-ignore
-      export const load = async ({ page }) => ({
-          props: {
-              key: page.path,
-          },
-      });
-  
-    import WelcomeMenu from "$lib/ui/WelcomeMenu.svelte";
-      /**
-       * @type {any}
-       */
-      export let key;
-    
-    
-  </script>
-  <WelcomeMenu />
-  <div class="container" >
- 
-   <SignupForm />
+  // https://www.npmjs.com/package/svelte-fa
+  import Footer from "$lib/Footer.svelte";
+  import Header from "$lib/Header.svelte";
+  import SignupForm from "./SignupForm.svelte";
 
-  </div>
-  <Footer />
+  // @ts-ignore
+  export const load = async ({ page }) => ({
+    props: {
+      key: page.path
+    }
+  });
+
+  import WelcomeMenu from "$lib/ui/WelcomeMenu.svelte";
+  /**
+   * @type {any}
+   */
+  export let key;
+</script>
+
+<!-- <WelcomeMenu /> -->
+<div class="container">
+  <SignupForm />
+</div>
+<!-- <Footer /> -->
